# replace fmt.Sprintf with string concatenations