this change significantly improves the speed of the parser,
especially on large documents