@xcoulon 2019-06-22
FYI, on a document that weights 65kb, I get the following speed improvements:

before:
```
$ time libasciidoc -s  ./tmp/large.adoc
libasciidoc -s ./tmp/large.adoc  0.22s user 0.04s system 118% cpu 0.222 total
```

after:
```
$ time libasciidoc -s  ./tmp/large.adoc
libasciidoc -s ./tmp/large.adoc  0.17s user 0.04s system 122% cpu 0.172 total
```