@xcoulon 2019-06-24
I've just updated the PR to apply the "root" changes at the `static_code.go` level and then regenerated all the files with `make all`. I did not realize that the initial commit of this PR included changes on a generated file 🤦‍♂️