@breml 2019-06-25
@xcoulon can you provide me a test case (grammar + document to parse), which allows me to verify the performance improvement?
An other interesting case would be a large JSON file, which can be tested with the existing JSON grammar.