@breml 2019-06-25
And I would like to have some Go benchmarks to verify the improvements. I quickly tested the PR against master with a ~600KB JSON file and the improvement looks like marginal (like 2.5% improvement).