@xcoulon 2019-06-26
@breml thanks for the review and the feedback. I'll get back to you with some proper benchmarks in my repo so we can compare my PR against `master` with some better measures than what I initially posted. I realize that this wasn't presented in a proper way, especially since measuring the time to execute the command from the shell involves way more that just the parser and it's a one-time call.
Sorry about that :/