@xcoulon 2019-06-30
hello @breml 

I've added some benchmark using the [Vert.x examples README](https://github.com/vert-x3/vertx-examples)  (11KB in this case) as the source to parse in the libasciidoc repository (see https://github.com/bytesparadise/libasciidoc/pull/370)

When running the bench with the parser generated with the `master` build of `pigeon` along with the `-optimize-parser` flag (`make generate-optimized`), I get the following results:

````
> ginkgo -focus "bench parser on 'vert.x examples' doc" pkg/parser
...
[MEASUREMENT]
parser benchmark
  Ran 50 samples:
  runtime:
    Fastest Time: 0.086s
    Slowest Time: 0.121s
    Average Time: 0.091s ± 0.005s
````

When I use my PR to make a new build of `pigeon` (make clean && make) and copy `./bin/pigeon` to `$GOPATH/bin`, then generate the parser again and run the same benchmark, I  get the following results:

````
...
[MEASUREMENT]
parser benchmark
  Ran 50 samples:
  runtime:
    Fastest Time: 0.068s
    Slowest Time: 0.098s
    Average Time: 0.073s ± 0.006s
````

As you can see, the avg duration went from 0.091s down to 0.073s

The file to parse is [here](https://raw.githubusercontent.com/bytesparadise/libasciidoc/master/test/bench/vertx-examples.adoc) and the grammar is [there](https://github.com/bytesparadise/libasciidoc/blob/master/pkg/parser/asciidoc-grammar.peg).

Please let me know if you need more information.