@xcoulon 2019-07-17
> @xcoulon Sorry for being slow in order to come back to you.
> 
> Have you tested/benchmarked, which of the 3 changes you made, added the greatest performance improvement? I investigated where [this](https://github.com/mna/pigeon/pull/86/files#diff-288fa6d159a3e33479b4e95cd813fd4fR224) line comes into play and to me it looks like it is only used for the message of this [panic](https://github.com/mna/pigeon/blob/master/builder/static_code.go#L1350). So to me it does not look like this change provides anything in regards to performance.
> 
> In regards to [preparing `lit.val` for `failAt`](https://github.com/mna/pigeon/pull/86/files#diff-288fa6d159a3e33479b4e95cd813fd4fR1239) it could be even faster to pre-calculate this value and add it to the `litMatcher` struct during generation of the parser.
> 
> And finally, if [`listJoin`](https://github.com/mna/pigeon/pull/86/files#diff-288fa6d159a3e33479b4e95cd813fd4fR872) is the bottleneck, we should test, if adding the case for `len(list) == 2` (or even 3 and 4) improves the performance even more.
> 
> Also I still would like to know, if `fmt.Sprintf` in `static_code.go:645` and `static_code.go:1161` have a measurable negative effect as well.
> 
> So the bottom line is: I like your discovery and I trust your benchmarks, but I would like to understand, where exactly the bottleneck is so we can think about additional measurements to improve the situation.

hello @breml and thanks for your response,

Ok, so I'll run distinct benchmarks for each single change (including the `listJoin` ones, where there are only 2 or 3 items) so we can measure their impact on the overall performances.
I'll post the results here.