@xcoulon 2019-07-20
hello @breml 

So, I ran some isolated benchs, each of them with a single change from this PR, and compared them to the `master` version. The benchs are executed on 3 different asciidoc document (see [here](https://github.com/bytesparadise/libasciidoc/blob/master/pkg/parser/bench_test.go#L45-L75)) and [here](https://github.com/bytesparadise/libasciidoc/blob/master/Makefile#L112-L116). 

The results are then compared with [benchstat](https://godoc.org/golang.org/x/perf/cmd/benchstat):

- when changing the `listJoin` function to support list of length 2 (not replacing the use of `fmt.Sprintf` for the `default` case):
```
name                 old time/op    new time/op    delta
Parser/1_line-8        57.2µs ± 2%    55.2µs ± 0%  -3.42%  (p=0.000 n=10+9)
Parser/10_lines-8       694µs ± 3%     687µs ± 1%    ~     (p=0.068 n=10+8)
Parser/vert.x_doc-8    71.2ms ± 1%    71.5ms ± 2%    ~     (p=0.237 n=8+10)

name                 old alloc/op   new alloc/op   delta
Parser/1_line-8        28.5kB ± 0%    28.5kB ± 0%    ~     (all equal)
Parser/10_lines-8       223kB ± 0%     223kB ± 0%  +0.00%  (p=0.002 n=10+9)
Parser/vert.x_doc-8    19.8MB ± 0%    19.8MB ± 0%    ~     (p=0.797 n=9+10)

name                 old allocs/op  new allocs/op  delta
Parser/1_line-8           417 ± 0%       417 ± 0%    ~     (all equal)
Parser/10_lines-8       6.39k ± 0%     6.39k ± 0%    ~     (all equal)
Parser/vert.x_doc-8      685k ± 0%      685k ± 0%    ~     (p=0.639 n=9+10)
```

- when changing the `listJoin` function to support list of length 2, 3 and 4 (still not replacing the use of `fmt.Sprintf` for the `default` case): 
```
name                 old time/op    new time/op    delta
Parser/1_line-8        57.2µs ± 2%    55.0µs ± 1%  -3.86%  (p=0.000 n=10+10)
Parser/10_lines-8       694µs ± 3%     684µs ± 2%  -1.55%  (p=0.001 n=10+9)
Parser/vert.x_doc-8    71.2ms ± 1%    71.7ms ± 2%    ~     (p=0.408 n=8+10)

name                 old alloc/op   new alloc/op   delta
Parser/1_line-8        28.5kB ± 0%    28.5kB ± 0%    ~     (all equal)
Parser/10_lines-8       223kB ± 0%     223kB ± 0%  +0.00%  (p=0.007 n=10+9)
Parser/vert.x_doc-8    19.8MB ± 0%    19.8MB ± 0%  +0.00%  (p=0.000 n=9+10)

name                 old allocs/op  new allocs/op  delta
Parser/1_line-8           417 ± 0%       417 ± 0%    ~     (all equal)
Parser/10_lines-8       6.39k ± 0%     6.39k ± 0%    ~     (all equal)
Parser/vert.x_doc-8      685k ± 0%      685k ± 0%  +0.00%  (p=0.005 n=9+10)
```

- when changing the `listJoin` function by replacing the use of `fmt.Sprintf` in the `default` case (but not supporting list of length 2, 3 or 4):
```
name                 old time/op    new time/op    delta
Parser/1_line-8        57.2µs ± 2%    55.2µs ± 1%  -3.53%  (p=0.000 n=10+9)
Parser/10_lines-8       694µs ± 3%     682µs ± 1%  -1.85%  (p=0.000 n=10+10)
Parser/vert.x_doc-8    71.2ms ± 1%    71.6ms ± 1%    ~     (p=0.173 n=8+10)

name                 old alloc/op   new alloc/op   delta
Parser/1_line-8        28.5kB ± 0%    28.5kB ± 0%    ~     (all equal)
Parser/10_lines-8       223kB ± 0%     223kB ± 0%  +0.00%  (p=0.001 n=10+9)
Parser/vert.x_doc-8    19.8MB ± 0%    19.8MB ± 0%    ~     (p=0.661 n=9+10)

name                 old allocs/op  new allocs/op  delta
Parser/1_line-8           417 ± 0%       417 ± 0%    ~     (all equal)
Parser/10_lines-8       6.39k ± 0%     6.39k ± 0%    ~     (all equal)
Parser/vert.x_doc-8      685k ± 0%      685k ± 0%    ~     (p=0.849 n=9+10)
```

- when removing the use of `fmt.Sprintf` in the `String` method:
```
name                 old time/op    new time/op    delta
Parser/1_line-8        57.2µs ± 2%    55.7µs ± 3%  -2.64%  (p=0.001 n=10+10)
Parser/10_lines-8       694µs ± 3%     682µs ± 0%  -1.80%  (p=0.000 n=10+9)
Parser/vert.x_doc-8    71.2ms ± 1%    71.3ms ± 1%    ~     (p=0.762 n=8+10)

name                 old alloc/op   new alloc/op   delta
Parser/1_line-8        28.5kB ± 0%    28.5kB ± 0%    ~     (all equal)
Parser/10_lines-8       223kB ± 0%     223kB ± 0%    ~     (p=0.284 n=10+10)
Parser/vert.x_doc-8    19.8MB ± 0%    19.8MB ± 0%    ~     (p=0.683 n=9+9)

name                 old allocs/op  new allocs/op  delta
Parser/1_line-8           417 ± 0%       417 ± 0%    ~     (all equal)
Parser/10_lines-8       6.39k ± 0%     6.39k ± 0%    ~     (all equal)
Parser/vert.x_doc-8      685k ± 0%      685k ± 0%    ~     (p=0.516 n=9+9)
```

- when removing the use of `fmt.Sprintf` in the `parseLitMatcher` method:
```
name                 old time/op    new time/op    delta
Parser/1_line-8        57.2µs ± 2%    49.7µs ± 1%  -13.14%  (p=0.000 n=10+10)
Parser/10_lines-8       694µs ± 3%     572µs ± 1%  -17.58%  (p=0.000 n=10+9)
Parser/vert.x_doc-8    71.2ms ± 1%    58.6ms ± 1%  -17.67%  (p=0.000 n=8+10)

name                 old alloc/op   new alloc/op   delta
Parser/1_line-8        28.5kB ± 0%    28.1kB ± 0%   -1.50%  (p=0.000 n=10+10)
Parser/10_lines-8       223kB ± 0%     216kB ± 0%   -3.19%  (p=0.000 n=10+8)
Parser/vert.x_doc-8    19.8MB ± 0%    19.0MB ± 0%   -3.98%  (p=0.000 n=9+10)

name                 old allocs/op  new allocs/op  delta
Parser/1_line-8           417 ± 0%       417 ± 0%     ~     (all equal)
Parser/10_lines-8       6.39k ± 0%     6.41k ± 0%   +0.28%  (p=0.000 n=10+10)
Parser/vert.x_doc-8      685k ± 0%      690k ± 0%   +0.76%  (p=0.000 n=9+8)
```

The main improvement in term of speed of execution comes from https://github.com/mna/pigeon/pull/86/files#diff-288fa6d159a3e33479b4e95cd813fd4fR1239. This change also sightly reduces the total of memory allocated (per operation), but has no real influence on the number of memory allocations (per operation).